tests/wpt/web-platform-tests/bluetooth/service/getCharacteristic/characteristic-found.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHealthThermometerDevice()
    .then(({device}) => device.gatt.getPrimaryService('health_thermometer'))
    .then(service => {
      return Promise.all([
        service.getCharacteristic(measurement_interval.alias),
        service.getCharacteristic(measurement_interval.name),
        service.getCharacteristic(measurement_interval.uuid)])
        .then(characteristics => {
          characteristics.forEach(characteristic => {
            assert_equals(
              characteristic.uuid, measurement_interval.uuid,
              'Characteristic UUID should be the same as requested UUID.');
            assert_equals(
              characteristic.service, service,
              'Characteristic service should be the same as service.');
          });
       });
    });
}, 'Request for characteristic. Should return right characteristic');
</script>
